Refactor to make the clang compiler happier, for cmake

+++ ORchestra/src/ORchestraEngine/DataSequence.cpp
@@ -1,16 +1,16 @@
 #include "DataSequence.h"
 
-DataSequence::DataSequence(const std::vector<StepData>& data) : mData(data)
+DataSequence::DataSequence(const std::vector<StepData> &data) : mData(data)
 {
 }
 
 StepData DataSequence::GetValue(const int index) const
 {
-    const int indexWrapped = index % mData.size();
-    return mData[indexWrapped];
+    const int indexWrapped = index % static_cast<int>(mData.size());
+    return mData[static_cast<unsigned long>(indexWrapped)];
 }
 
 void DataSequence::SetValue(int index, StepData value)
 {
-    mData[index] = value;
+    mData[static_cast<unsigned long>(index)] = value;
 }
